# Conflicts:
#	databases/schedule.db
#	modules/web_routes.py
#	styles/base.css
#	styles/colors/black.css
#	styles/colors/blue.css
#	styles/colors/green.css
#	styles/colors/purple.css
#	styles/colors/red.css
#	styles/colors/sky.css
#	styles/colors/white.css
#	templates/about.html

+++ modules/schedule_access.py
@@ -11,9 +11,14 @@ class User(object):
     __month_list = ['январь', 'февраль', 'март', 'апрель', 'май', 'июнь', 'июль', 'август', 'сентябрь', 'октябрь',
                     'ноябрь', 'декабрь', ]
 
-    def __init__(self, log, psw=None):
+    def __init__(self, log, email, psw=None):
         User.__authorization = True
-        self.log = log
+        if inj_check(log) and inj_check(email):
+            self.log = log
+            self.email = email
+        else:
+            self.log = 'Test_user'
+            self.email = 'test@test.ru'
         if not User.check_log(self):
             User.__add_user(self, psw)
         User.__cur.execute("SELECT theme, color FROM users WHERE login=?", (self.log,))
@@ -48,8 +53,8 @@ class User(object):
 
     def __add_user(self, psw):
         psw = User.__gen(psw)
-        User.__cur.execute("INSERT INTO users (login, psw, theme, color) VALUES (?, ?, ?, ?)",
-                           (self.log, psw, 'light', 'blue'))
+        User.__cur.execute("INSERT INTO users (login, psw, email, theme, color, avatar) VALUES (?, ?, ?, ?, ?, ?)",
+                           (self.log, psw, self.email, 'light', 'blue', False))
         User.__conn.commit()
         User.__cur.executescript(f"""
             CREATE TABLE IF NOT EXISTS month_{self.log} (digit INTEGER, month VARCHAR(30), task VARCHAR(1000));
@@ -66,42 +71,57 @@ class User(object):
         return User.__check(User.__cur.fetchone(), psw)
 
     def change_log(self, log):
-        User.__cur.execute("UPDATE users SET login=? WHERE login=?", (log, self.log))
-        User.__cur.executescript(f"""
-            ALTER TABLE month_{self.log} RENAME TO month_{log};
-            ALTER TABLE day_{self.log} RENAME TO day_{log};
-            ALTER TABLE list_{self.log} RENAME TO list_{log}
-        """)
-        User.__conn.commit()
-        self.log = log
+        if inj_check(log):
+            User.__cur.execute("UPDATE users SET login=? WHERE login=?", (log, self.log))
+            User.__cur.executescript(f"""
+                ALTER TABLE month_{self.log} RENAME TO month_{log};
+                ALTER TABLE day_{self.log} RENAME TO day_{log};
+                ALTER TABLE list_{self.log} RENAME TO list_{log}
+            """)
+            User.__conn.commit()
+            self.log = log
 
     def change_pass(self, psw):
         psw = User.__gen(psw)
         User.__cur.execute("UPDATE users SET psw=? WHERE login=?", (psw, self.log))
         User.__conn.commit()
 
+    def change_email(self, email):
+        if inj_check(email):
+            User.__cur.execute("UPDATE users SET email=? WHERE login=?", (email, self.log))
+            User.__conn.commit()
+            self.email = email
+
     def change_theme(self, theme):
-        self.theme = theme
-        User.__cur.execute("UPDATE users SET theme=?, color=? WHERE login=?", (*self.theme, self.log))
-        User.__conn.commit()
+        if inj_check(theme[0]) and inj_check(theme[1]):
+            self.theme = theme
+            User.__cur.execute("UPDATE users SET theme=?, color=? WHERE login=?", (*self.theme, self.log))
+            User.__conn.commit()
+
+    def change_avatar(self, change):
+        if change is not None:
+            User.__cur.execute("UPDATE users SET avatar=? WHERE login=?", (change, self.log))
+            User.__conn.commit()
+        User.__cur.execute("SELECT (avatar) FROM users WHERE login=?", (self.log,))
+        return User.__cur.fetchone()[0]
 
     def del_user(self):
         User.__cur.execute("DELETE FROM users WHERE login=?", (self.log,))
         User.__conn.commit()
         User.__cur.executescript(f"""
-			DROP TABLE IF EXISTS month_{self.log};
-        	DROP TABLE IF EXISTS day_{self.log};
-        	DROP TABLE IF EXISTS list_{self.log}
-		""")
+            DROP TABLE IF EXISTS month_{self.log};
+            DROP TABLE IF EXISTS day_{self.log};
+            DROP TABLE IF EXISTS list_{self.log}
+        """)
 
     def add_list(self, name, task):
-        if (name, task) not in self.lists:
+        if (name, task) not in self.lists and inj_check(name) and inj_check(task):
             User.__cur.execute(f"INSERT INTO list_{self.log} (name, task) VALUES (?, ?)", (name, task))
             User.__conn.commit()
             self.lists = User.__ret_list(self)
 
     def add_day(self, hour, minute, task):
-        if (hour, minute, task) not in self.day:
+        if (hour, minute, task) not in self.day and inj_check(task):
             User.__cur.execute(f"INSERT INTO day_{self.log} (hour, minute, task) VALUES (?, ?, ?)",
                                (hour, minute, task))
             User.__conn.commit()
@@ -109,26 +129,26 @@ class User(object):
 
     def add_month(self, digit, month, task):
         month = month.lower()
-        if (digit, month, task) not in self.month:
+        if (digit, month, task) not in self.month and inj_check(task):
             User.__cur.execute(f"INSERT INTO month_{self.log} (digit, month, task) VALUES (?, ?, ?)",
                                (digit, month, task))
             User.__conn.commit()
             self.month = User.__ret_month(self)
 
     def del_list(self, name):
-        if (name,) in self.lists:
+        if (name,) in self.lists and inj_check(name):
             User.__cur.execute(f"DELETE FROM list_{self.log} WHERE name = ?", (name,))
             User.__conn.commit()
             self.lists.pop(name)
 
     def del_list_task(self, name, task):
-        if task in self.lists[name] and task != 0:
+        if task in self.lists[name] and task != 0 and inj_check(name) and inj_check(task):
             User.__cur.execute(f"DELETE FROM list_{self.log} WHERE name = ? AND task = ?", (name, task))
             User.__conn.commit()
             self.lists[name].remove(task)
 
     def del_day(self, hour, minute, task):
-        if (hour, minute, task) in self.day:
+        if (hour, minute, task) in self.day and inj_check(task):
             User.__cur.execute(f"DELETE FROM day_{self.log} WHERE hour = ? AND minute = ? AND task = ?",
                                (hour, minute, task))
             User.__conn.commit()
@@ -136,14 +156,14 @@ class User(object):
 
     def del_month(self, digit, month, task):
         month = month.lower()
-        if (digit, month, task) in self.month:
+        if (digit, month, task) in self.month and inj_check(task):
             User.__cur.execute(f"DELETE FROM month_{self.log} WHERE digit = ? AND month = ? AND task = ?",
                                (digit, month, task))
             User.__conn.commit()
             self.month.remove((digit, month, task))
 
     @staticmethod
-    def __erase():
+    def _erase():
         """Стирание всех пользователей"""
         User.__cur.execute("SELECT login FROM users")
         log_list = User.__cur.fetchall()
@@ -166,18 +186,26 @@ def inj_check(req):
     return True
 
 # Создание таблицы-----------------------------------------------------------------------------------------
-# __cur.execute("CREATE TABLE IF NOT EXISTS users (login VARCHAR(200), psw VARCHAR(200), theme VARCHAR(30), color VARCHAR(30))")
+# __cur.execute("""CREATE TABLE IF NOT EXISTS users
+#                   (login VARCHAR(200), psw VARCHAR(200),
+#                   email VARCHAR(200), theme VARCHAR(30), color VARCHAR(30), avatar BOOLEAN)""")
 # -----------------------------------------------------------------------------------------------------------
 # Тесты (Артем и Дима(ахах, норм вписался)) print(inj_check('adsfghdffdsfgfdf')) User._erase()
-# now_user = User("T1MON", 'kdfjdkffj') now_user2 = User("T1MON", 'asdfss') now_user1 = User("TKACH", 'sfdsd') print(
+# now_user = User("T1MON", 'T1MON@yandex.ru', 'kdfjdkffj')
+# now_user2 = User("T1MON", 'asdfss') now_user1 = User("TKACH", 'sfdsd') print(
 # now_user.log) now_user.change_log('ATTILENE') print(now_user.log) print(now_user.day) print(now_user.month) print(
 # now_user.lists)
+# print(now_user.log)
+# print(now_user.email)
+# now_user.change_email('qwerty@mail.ru')
+# print(now_user.email)
+# print(now_user.change_avatar(False))
 # now_user.add_month(23, 'январь', 'dfjfkdjf')
 # now_user.add_month(24, 'январь', 'dfjfkdjf')
 # now_user.add_month(25, 'январь', 'dfjfkdjf')
-# now_user.add_list(1, 'dfjfkdjfdsfdgf')
-# now_user.add_list(2, 'dfjf')
-# now_user.add_list(1, 'dfjfkdjfdsfdgfasdfgdhfgjhjrsdv')
+# now_user.add_list('1', 'dfjfkdjfdsfdgf')
+# now_user.add_list('2', 'dfjf')
+# now_user.add_list('1', 'dfjfkdjfdsfdgfasdfgdhfgjhjrsdv')
 # now_user.add_day(22, 33, 'jdfkjf')
 # now_user.add_day(24, 33, 'jdfkjf')
 # now_user.add_day(23, 33, 'jdfkjf')
